Refactor bootloader configuration file changes
We basically need to make the same changes to both the pxelinux and GRUB
configuration files - i.e., adding kernel parameters.  Today we do this
using two copies of mostly identical code. Recent changes that required
kernel parameter manipulation have shown that it is easy to update one
and forget to update the other. This change aims to consolidate that code
to prevent that.

We centralize all of the kernel parameter logic in the new function
setup_kernel_params() that can operate on any subclass of BootloaderConfig.
We provide such subclasses for GRUB and pxelinux, which have the internal
knowledge for creating and editing these files (which isn't much). Since
setup_kernel_params() is called for both files, any kernel parameter logic
changes should get applied to both equivalently.